Organize and extend exports
