+++ build.preset.ts
@@ -2,7 +2,7 @@ import { definePreset } from 'unbuild';
 
 export default definePreset({
 	clean: true,
-	declaration: true,
+	declaration: 'node16',
 	rollup: {
 		esbuild: {
 			target: 'es2022',
