+++ packages/core/scripts/schema.ts
@@ -1,7 +1,7 @@
 import { writeFileSync } from 'node:fs';
 import { resolve } from 'node:path';
 import { zodToJsonSchema } from 'zod-to-json-schema';
-import { LunariaConfigSchema } from '../src/config/schemas.js';
+import { LunariaConfigSchema } from '../src/config/index.js';
 
 const schema = JSON.stringify(zodToJsonSchema(LunariaConfigSchema, 'LunariaConfigSchema'), null, 2);
