+++ packages/core/src/cli/sync/index.ts
@@ -1,8 +1,13 @@
 import { readFileSync } from 'node:fs';
 import { resolve } from 'node:path';
+import {
+	loadConfig,
+	writeConfig,
+	type LunariaConfig,
+	type LunariaUserConfig,
+} from '../../config/index.js';
+import { bold, confirm, error, handleCancel, highlight, sync as s, select } from '../console.js';
 import { getFormattedTime } from '../helpers.js';
-import { bold, error, highlight, sync as s } from '../messages.js';
-import { handleCancel, select } from '../prompts.js';
 import type { PackageJson, SyncOptions } from '../types.js';
 
 /** Packages that we support sync with. */
@@ -88,4 +93,66 @@ function loadPackageJson() {
 		console.error(error('Failed to load your package.json\n'));
 		throw e;
 	}
+}
+
+export async function updateConfig(
+	configPath: string,
+	defaultLocale: LunariaConfig['defaultLocale'] | undefined,
+	locales: LunariaConfig['locales'] | undefined,
+	file: LunariaUserConfig['files'][number],
+	skip: boolean
+) {
+	const { rawUserConfig: config } = await loadConfig(configPath);
+
+	if (defaultLocale) {
+		let answer: boolean = true;
+		if (!skip) {
+			const updateDefaultLocale = await confirm({
+				message: s('Update defaultLocale?'),
+				initialValue: true,
+			});
+
+			handleCancel(updateDefaultLocale);
+
+			answer = Boolean(updateDefaultLocale);
+		}
+
+		if (answer || skip) config.defaultLocale = defaultLocale;
+	}
+
+	if (locales) {
+		let answer: boolean = true;
+		if (!skip) {
+			const updateLocales = await confirm({
+				message: s('Update locales?'),
+				initialValue: true,
+			});
+
+			handleCancel(updateLocales);
+
+			answer = Boolean(updateLocales);
+		}
+
+		if (answer || skip) config.locales = locales;
+	}
+
+	if (file) {
+		let answer: boolean = true;
+		if (!skip) {
+			const updateFiles = await confirm({
+				message: s('Update files?'),
+				initialValue: true,
+			});
+
+			handleCancel(updateFiles);
+
+			answer = Boolean(updateFiles);
+		}
+
+		const otherFiles = config.files?.filter((f) => f.location !== file.location) ?? [];
+
+		if (answer || skip) config.files = [file, ...otherFiles];
+	}
+
+	writeConfig(configPath, config);
 }
