+++ packages/core/src/config/index.ts
@@ -1,7 +1,7 @@
-import { readFileSync } from 'node:fs';
+import { readFileSync, writeFileSync } from 'node:fs';
 import { resolve } from 'node:path';
 import { fromZodError } from 'zod-validation-error';
-import { error } from '../cli/messages.js';
+import { error } from '../cli/console.js';
 import { loadWithJiti } from '../utils.js';
 import {
 	LunariaConfigSchema,
@@ -9,6 +9,8 @@ import {
 	type LunariaUserConfig,
 	type LunariaUserRendererConfig,
 } from './schemas.js';
+
+export * from './schemas.js';
 
 const fromZodErrorOptions = {
 	prefix: '- ',
@@ -21,11 +23,11 @@ export async function loadConfig(path: string) {
 
 	if (/\.json$/.test(resolvedPath)) {
 		try {
-			const rawConfig = JSON.parse(readFileSync(resolvedPath, 'utf-8'));
-			const userConfig = validateConfig(rawConfig);
+			const rawUserConfig = JSON.parse(readFileSync(resolvedPath, 'utf-8')) as LunariaUserConfig;
+			const userConfig = validateConfig(rawUserConfig);
 			const rendererConfig = await loadRendererConfig(userConfig.renderer);
 
-			return { userConfig, rendererConfig };
+			return { rawUserConfig, userConfig, rendererConfig };
 		} catch (e) {
 			console.error(error('Failed to load Lunaria config\n'));
 			throw e;
@@ -79,5 +81,22 @@ export function validateRendererConfig(config: LunariaUserRendererConfig) {
 	const validationError = fromZodError(parsedConfig.error, fromZodErrorOptions);
 
 	console.error(error('Invalid Lunaria renderer config:\n' + validationError));
+	process.exit(1);
+}
+
+export function writeConfig(path: string, config: LunariaUserConfig) {
+	const resolvedPath = resolve(path);
+
+	if (/\.json$/.test(resolvedPath)) {
+		try {
+			const configJSON = JSON.stringify(config, null, 2);
+			writeFileSync(resolvedPath, configJSON);
+		} catch (e) {
+			console.error(error('Failed to write Lunaria config\n'));
+			throw e;
+		}
+	}
+
+	console.error(error('Invalid Lunaria config extension, expected .json'));
 	process.exit(1);
 }
